app/requirements.txt
Flask==2.3.4
python-ldap==3.5.0
